requirements.txt
fastapi
uvicorn
streamlit
pdfplumber
python-docx
pandas
python-pptx
openpyxl
python-multipart
chromadb
openai
python-dotenv
pytesseract
groq
